Revert "Refactor: 커뮤니티 상단바 수정"
