+++ app/build.gradle.kts
@@ -74,7 +74,6 @@ dependencies {
     implementation("androidx.preference:preference:1.2.1")
     implementation ("com.google.android.material:material:1.12.0")
 
-    implementation("androidx.recyclerview:recyclerview:1.3.2")
 
     implementation("com.squareup.retrofit2:retrofit:2.0.0")
     implementation("com.google.code.gson:gson:2.8.9")
