--- app/src/main/java/com/example/walking_hadang/ui/community/PostDetailFragment.kt
@@ -1,4 +1,0 @@
-package com.example.walking_hadang.ui.community
-
-class PostDetailFragment {
-}
