--- app/src/main/res/color/cat_btn_text_color.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:state_selected="true" android:color="#FFFFFF"/>
-    <item android:state_pressed="true" android:color="#000000"/>
-    <item android:color="#000000"/>
-</selector>
