--- app/src/main/res/drawable/bg_board_title.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-
-    <!-- 배경색 흰색 -->
-    <solid android:color="#FFFFFF" />
-
-    <!-- 모서리 radius -->
-    <corners android:radius="10dp" />
-
-    <!-- 테두리 -->
-    <stroke
-        android:width="1dp"
-        android:color="#0F6972" />
-</shape>
