--- app/src/main/res/drawable/bg_new_badge.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-    <solid android:color="#bf4040" />
-    <corners android:radius="6dp"/>
-    <padding
-        android:left="6dp"
-        android:top="2dp"
-        android:right="6dp"
-        android:bottom="2dp"/>
-</shape>
