--- app/src/main/res/drawable/btn_category.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <!-- 선택됨 -->
-    <item android:state_selected="true">
-        <shape android:shape="rectangle">
-            <solid android:color="#14000000"/>
-            <corners android:radius="30dp"/>
-        </shape>
-    </item>
-    <!-- 눌림 -->
-    <item android:state_pressed="true">
-        <shape android:shape="rectangle">
-            <solid android:color="#14000000"/>
-            <corners android:radius="30dp"/>
-            <stroke android:width="1dp" android:color="#D3A33A"/>
-        </shape>
-    </item>
-    <!-- 기본 -->
-    <item>
-        <shape android:shape="rectangle">
-            <solid android:color="#FFFFFF"/>
-            <corners android:radius="30dp"/>
-            <stroke android:width="1dp" android:color="#E0E0E0"/>
-        </shape>
-    </item>
-</selector>
