--- app/src/main/res/layout/fragment_exercise_board.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/exerciseRoot"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="#FFFFFF">
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rvBoardPosts"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:clipToPadding="false"
-        android:paddingTop="8dp"
-        android:paddingBottom="96dp"
-        android:paddingStart="12dp"
-        android:paddingEnd="12dp"
-        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager" />
-
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:id="@+id/fabAddPost"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_margin="16dp"
-        android:contentDescription="@string/app_name"
-        app:srcCompat="@drawable/add"
-        app:tint="#FFFFFF"
-        app:backgroundTint="#9CC5B5"
-        app:layout_anchor="@id/rvBoardPosts"
-        app:layout_anchorGravity="bottom|end" />
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
