--- app/src/main/res/layout/fragment_home_hot.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:padding="8dp"
-    android:background="#FFF1C7">
-
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rvHot"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:paddingTop="8dp"
-        android:clipToPadding="false" />
-</LinearLayout>
