--- app/src/main/res/layout/fragment_sugar_board.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/sugarRoot"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="#FFFFFF">
-
-    <!-- 목록 -->
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rvSugarPosts"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:clipToPadding="false"
-        android:paddingTop="8dp"
-        android:paddingBottom="96dp"
-        android:paddingStart="12dp"
-        android:paddingEnd="12dp"
-        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager" />
-
-    <!-- + 글쓰기 버튼 : CoordinatorLayout의 직계 자식으로 둬야 항상 겹쳐 보임 -->
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:id="@+id/fabAddPost"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:contentDescription="@string/app_name"
-        android:layout_margin="16dp"
-        app:srcCompat="@drawable/add"
-        app:tint="#FFFFFF"
-        app:backgroundTint="#9CC5B5"
-        app:layout_anchor="@id/rvSugarPosts"
-        app:layout_anchorGravity="bottom|end" />
-
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
