--- app/src/main/res/layout/item_comment.xml
@@ -1,107 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="vertical"
-    android:padding="12dp">
-
-    <!-- 프로필 + 닉네임/시간/좋아요 -->
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="center_vertical">
-
-        <!-- 프로필 이미지 -->
-        <ImageView
-            android:layout_width="28dp"
-            android:layout_height="28dp"
-            android:src="@drawable/ic_profile2"/>
-
-        <!-- 닉네임 + 시간 -->
-        <LinearLayout
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="8dp"
-            android:layout_weight="1"
-            android:orientation="vertical">
-
-            <TextView
-                android:id="@+id/tvNick"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="익명 1"
-                android:textStyle="bold" />
-
-
-        </LinearLayout>
-
-        <!-- 좋아요/싫어요 -->
-        <LinearLayout
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:gravity="center_vertical"
-            android:orientation="horizontal">
-
-            <ImageView
-                android:layout_width="16dp"
-                android:layout_height="16dp"
-                android:src="@drawable/ic_like" />
-
-            <TextView
-                android:id="@+id/tvCommentLike"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginLeft="2dp"
-                android:layout_marginEnd="12dp"
-                android:text="3" />
-
-            <ImageView
-                android:layout_width="16dp"
-                android:layout_height="16dp"
-                android:src="@drawable/ic_comment" />
-
-            <TextView
-                android:id="@+id/tvCommentDislike"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginLeft="2dp"
-                android:text="0" />
-        </LinearLayout>
-    </LinearLayout>
-
-    <!-- 댓글 내용 -->
-    <TextView
-        android:id="@+id/tvCommentContent"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="6dp"
-        android:text="저 러닝 시작할때 나이키 에어 줌 페가수스 39 샀는데 꽤 괜찮았어요 가격대도 ㅋㅋ" />
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:gravity="center_vertical"
-        android:orientation="horizontal">
-
-        <TextView
-            android:id="@+id/tvTime"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="08/19 13:50"
-            android:textSize="12sp"
-            android:textColor="#999999"/>
-
-        <ImageView
-            android:layout_width="11dp"
-            android:layout_height="11dp"
-            android:layout_marginLeft="4dp"
-            android:src="@drawable/ic_like" />
-    </LinearLayout>
-
-    <!-- 구분선 -->
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="1dp"
-        android:layout_marginTop="12dp"
-        android:background="#EEEEEE"/>
-</LinearLayout>
