--- app/src/main/res/layout/item_community_post.xml
@@ -1,128 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="vertical"
-    android:background="#FFFFFF"
-    android:padding="10dp"
-    android:layout_marginBottom="4dp"
-    android:elevation="2dp">
-
-    <!-- 게시판 배지 -->
-    <TextView
-        android:id="@+id/tvBadge"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="혈당 공유"
-        android:textSize="12sp"
-        android:textStyle="bold"
-        android:textColor="@android:color/holo_red_dark"
-        android:background="@android:color/darker_gray"
-        android:paddingHorizontal="6dp"
-        android:paddingVertical="2dp"
-        android:layout_marginBottom="4dp"
-        android:visibility="gone" />
-
-    <!-- 글 제목 + NEW 뱃지 -->
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal">
-
-        <!-- 글 제목 -->
-        <TextView
-            android:id="@+id/tvPostTitle"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:text="고혈압 관리 어떻게들 하시나요ㅠㅠ??"
-            android:textSize="16sp"
-            android:textStyle="bold"
-            android:textColor="#000000"/>
-
-        <!-- NEW 뱃지 -->
-        <TextView
-            android:id="@+id/tvNew"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="N"
-            android:textStyle="bold"
-            android:textColor="#FFFFFF"
-            android:textSize="11sp"
-            android:background="@drawable/bg_new_badge"
-            android:paddingLeft="6dp"
-            android:paddingRight="6dp"
-            android:paddingTop="2dp"
-            android:paddingBottom="2dp"
-            android:layout_marginStart="4dp"
-            android:visibility="gone"/>
-    </LinearLayout>
-
-    <!-- 글 내용 미리보기 -->
-    <TextView
-        android:id="@+id/tvPostPreview"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="저녁에 운동도 꾸준히 하는데 너무 힘드네요..."
-        android:textSize="14sp"
-        android:textColor="#444444"
-        android:layout_marginTop="4dp"/>
-
-    <!-- 좋아요/댓글/시간 -->
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:layout_marginTop="6dp">
-
-        <LinearLayout
-            android:id="@+id/reactionBar"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="16dp"
-            android:gravity="center_vertical"
-            android:orientation="horizontal">
-
-            <ImageButton
-                android:id="@+id/btnLike"
-                android:layout_width="12dp"
-                android:layout_height="12dp"
-                android:background="?attr/selectableItemBackgroundBorderless"
-                android:src="@drawable/ic_like" />
-
-            <TextView
-                android:id="@+id/tvLikeCount"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginEnd="12dp"
-                android:text="0" />
-
-
-
-            <ImageView
-                android:layout_width="12dp"
-                android:layout_height="12dp"
-                android:src="@drawable/ic_comment"/>
-
-            <TextView
-                android:id="@+id/tvCommentCount"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="4dp"
-                android:text="0" />
-
-            <TextView
-                android:id="@+id/tvTime"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginLeft="4dp"
-                android:text="2시간 전"
-                android:textSize="12sp"
-                android:textColor="#888888"/>
-
-        </LinearLayout>
-
-
-    </LinearLayout>
-
-</LinearLayout>
