--- app/src/main/res/menu/menu_community_toolbar.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-    <item
-        android:id="@+id/menu_search"
-        android:icon="@drawable/ic_search"
-        android:title="Search"
-        app:showAsAction="always" />
-    <item
-        android:id="@+id/menu_bell"
-        android:icon="@drawable/ic_notify"
-        android:title="Profile"
-        app:showAsAction="always" />
-
-    <item
-    android:id="@+id/menu_profile"
-    android:icon="@drawable/ic_profile"
-    android:title="Profile"
-    app:showAsAction="always" />
-</menu>
