--- app/src/main/res/menu/menu_post_other.xml
@@ -1,6 +1,0 @@
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:id="@+id/action_report" android:title="신고"  android:icon="@drawable/ic_report" />
-    <item android:id="@+id/action_block"  android:title="차단"  android:icon="@drawable/ic_block"  />
-    <item android:id="@+id/action_share"  android:title="공유"  android:icon="@drawable/ic_share"  />
-</menu>
-
