--- app/src/main/res/menu/menu_post_owner.xml
@@ -1,6 +1,0 @@
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:id="@+id/action_edit"   android:title="수정"  android:icon="@drawable/ic_edit"    />
-    <item android:id="@+id/action_delete" android:title="삭제"  android:icon="@drawable/ic_delete"  />
-    <item android:id="@+id/action_share"  android:title="공유"  android:icon="@drawable/ic_share"   />
-</menu>
-
