+++ app/src/main/res/menu/menu_toolbar.xml
@@ -2,6 +2,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
+
     <item
     android:id="@+id/menu_profile"
     android:icon="@drawable/ic_profile"
