--- app/src/main/res/menu/menu_write_post.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-    <item
-        android:id="@+id/action_done"
-        android:title="완료"/>
-</menu>
